resolves #5: Freeze on new part in new document
Loading a NULL part does not update the wemedpanel viewer. Consequently
the later attempt to fetch its contents blocks indefinitely. Change the
mime part content loading logic so that a NULL return won't happen and
an empty string is returned instead.

That fixes the general case, but if the part is of type text/html then
the same problem occurs when trying to fetch the contents since
webkit_web_view_load_bytes refuses to load a GBytes with zero length.
In this case, pass the NULL terminator as part of the GBytes length.