Edge-case fix for simultaneous webkit and sourceview widgets
The following sequence of actions causes both the WebkitWebView and
the GtkSourceView to be visible simultaneously:

- New Document
- Add Empty Part
- Add Empty Part
- Change part type to text/html
- Select the other empty part

This is because the header type change causes a reload of the part
at the same time as the other empty part is supposed to be loaded.
The load is asynchronous, so the webkit load of the html part finishes
after the load of the sourceview part, which makes the webkit view
visible.

Fix by removing the reloading of the part when its headers are changed.
This only happened when navigating away from the part so it's of
questionable utility.